requirements.txt
alabaster==0.7.11
Babel==2.6.0
certifi==2018.4.16
chardet==3.0.4
click==6.7
docutils==0.14
Flask==0.11.1
Flask-Admin==1.4.1
Flask-Cache==0.13.1
Flask-Cors==3.0.3
Flask-Login==0.2.11
flask-swagger==0.2.13
Flask-WTF==0.14
gunicorn==19.7.1
idna==2.7
imagesize==1.0.0
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
numpy==1.13.3
numpydoc==0.7.0
packaging==17.1
pandas==0.20.3
Pygments==2.2.0
PyJWT==1.6.4
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.5
PyYAML==3.13
requests==2.19.1
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.7.5
sphinxcontrib-websupport==1.1.0
urllib3==1.23
Werkzeug==0.14.1
WTForms==2.2.1
